+++ libs/common/src/lib/validator-constraints/interfaces/interfaces.ts
@@ -1,0 +1,6 @@
+import { DataSource, Type } from '@prisma/client';
+
+export interface ImportDataLike {
+  activities?: { dataSource?: DataSource; symbol?: string; type?: Type }[];
+  assetProfiles?: { dataSource?: DataSource; symbol?: string }[];
+}